if input relative patterns, output relative paths
Previously, any globbing operation resulted in absolute paths regardless
of whether or not the patterns were relative. This was due to
special-case behavior for handling volume-relative patterns on windows.

This change defines platform-dependent functions which determine the
scope, affording windows the ability to handle volume-relative paths.

The advantage of this is that now if relative paths are input, matched
paths are output as relative as well.