adds error reporting to Pattern and Glob
Instead of silently treating certain invalid patterns literally,
potentially surprising users, this change adds error reporting.

Specifically, the `Pattern::new` method now yields a
`Result<Pattern, Error>`, much like the `Regex::new` method.
When an invalid pattern is encountered, an `Error` type is returned
which contains an error message along with the approximate character
position of the error.

This error handling also bubbles up to the `glob` and `glob_with`
functions which now also return a `Result`.

The documentation is updated to reflect these changes, as well as
describe which patterns are invalid, which as of now include:

* an unclosed opening bracket `[`. If they want a literal one, they
  can escape it.

* a sequence of more than 2 wildcards, e.g. `***`. wildcards must
  either be regular `*` or recursive `**`, literal sequences of
  more than 2 asterisks should be escaped using the `escape` method

* a recursive wildcard that does not entirely form a single path
  component. e.g. a/**b/c or a/b**c/d. it must be e.g. a/**/b

This also adds a `String` field to the `Pattern` which is used to store
the original glob pattern, like the `Regex` type does, along with an
`as_str()` method to yield a slice to this field, as well as
`fmt::String` and `fmt::Show` implementations which show this
original glob pattern.