enables IO error propagation
This allows previously-repressed IO errors to propagate during
iteration. In particular, the Iterator `Item` is now a `Result` type
which represents either a matched `Path` or a `GlobError`, which is
simply a combination of an `IoError` and the related `Path` (for
context). Currently, the only potential for errors is during the
call to `readdir` and the errors are likely to be due to the lack of
permissions. An emitted error in this case essentially means that the
given path matched or partially matched, but its contents could not be
read to determine if they match.

It's important to note that this is an error for a given iteration, that
is, it doesn't completely halt the iteration process. That is, errors
are still returned as `Some()`. This enables the user to simply skip
the unreadable paths by using something like filter_map(Result::ok),
essentially replicating the previous behavior.