Fix incorrect setting of credential.helper
While the fetching code used a credential helper
based on an environment variable, the push code
still used the old one that causes a concurrency bug
when multiple requests are handled at the same time.
Now the credential helper is only set in one place
and is always using the environment variable to pass the
credentials.