Remove backward_maps
Turns out they are not really needed:
During reverse filter (push) only the very first's
commit of the to be pushed sequence can possibly
be in that table. All others have to be calculated
by traversing the new commits, and the results of
that where never kept in the first place.
For the very first the parent can be found by searching
the target branch in forward topological order, which will
almost always be a very short search as long as the new
commits are not based on an ancient version of the target.