Rework josh documentation
Reorder documentation to direct newcomers to the best introductory
information first.

Rework proxy documentation to (hopefully) better explain how the
josh-proxy works and the syntax of the URLs that users will need to
write to use the proxy.

Add to the filters documentation to explain the caveats of repository
naming and importance of filter order to teach newcomers how to
explicitly name their repositories, and also why filter order matters so
much.