Fix multiple edits of workspace file
When a workspace file was edited from a workspace more than
once, amended and pushed to the same target ref
(only really possible with magic refs like gerrit refs/for/*)
The commit being pushed already is a merge commit (if additional
paths where included) and the parent introduced through the workspace
change should not end up in the unfiltered history.
josh-proxy was rejecting such a second push because the additional
parent contained unrelated history. However that is not helpful
for the users workflow.
So the new behaviour is remove unknown orphans and their descendants
from the push unless the "josh-merge" option was specified.