Back to sparse cache
Once again, only store selected commits in the persistent cache
but also include merge and orphan commits in the cache.
This seems to work well on previously problematic scenarios and
also gives most of the space saving advantage.
Also remove the /clear_cache endpoint because it does not really
work.