Add counter for proxy integration test to prevent endless loop(s)
In some development environments the josh-proxy integration tests end up in an endless loop.
The root cause is not yet identified. Therefore an additional counter have been
added, to make sure the developer gets a failing test as feedback, instead of
getting stuck in an endless loop for no obvious reason.