Refactor: define git commands as slices, not strings
Slices will be passed directly to the system call; strings will be
interpreted by shell, where word splitting will occur. No word splitting
and no shell is better for security, and more convenient because one
level of quote escaping goes away. This is needed for SSH support
because of security implications, as well as because SSH commands will
be more complicated

commit-id:1c53a1f1