fix: allow building josh from a tarball download
The `VERSION` compile-time constant expects `git` and the josh git
repository to be present at build time, however this is not the case
when building from (for example) a tarball downloaded from GitHub.

The latter is the preferred method of building packages in Nix and
other distribution package managers, where fetching the entire git
history for the package build is not required and the revision is
known anyways.

This change falls back to using the `JOSH_VERSION` environment
variable if it is present at compile-time, or ultimately to the string
`unknown` if no version can be determined.

Note that the explicit match is required because methods like
`Option<T>::unwrap_or` are not currently `const` in stable Rust, but
const destructuring of sum types is.