Fix optimisation bug for compose filter
The optimiser incorrectly transformed :[:/y,:/x/y] into :[:/,:/x]:/y
This is not correct because :/y is more restrictive than :/ and will
therefore not match all the files that :/ does.
Due to the compose filter guaranteeing uniqueness of mappings
the led to files missing from the first filter.

Change: fix-opt-bug