Fix generated merge commit on workspace redirect
The :workspace filter compares the workspace mapping
between commits and creates a merge commit with the
history of newly mapped paths.

This did not work correctly in case the parents workspace
mapping was a redirect, in which case it considered all
paths as newly mapped even though they are already part
of the history.

Now the redirect is taken into account when parting the
parents workspace definition and we get correct history.

Change: fix-redirect-merge