Make rev() detect all parents of the given commits
The `:rev()` filter used to only apply filters if the specified
revisions where encountered during traversal.
However, this meant that if an ancestor commit is reachable by
multiple paths in the history and some include the specified
revision and some don't, the ancestor has multiple corresponding
filtered commits.
This resulted in duplicated commits in those more complicated
histories and caused "non roundtrip" issues.
Now the rev filter will compute if a commit is reachable from the
specifed commit in any way and apply if so.

Change: rev-all-parents
Co-authored-by: Ralf Jung <post@ralfj.de>