Refactor: drop gitoxide from josh-core
While gitoxide provides a well designed set of APIs, it's more oriented
towards implementing a git client than a server, and it lacks some
high-level APIs like treebuilder etc. Having it alongside the regular
git2 implementation works but creates additional I/O pressure which has
performance implications in scenarios when a lot of transactions are
opened (graphql). It would still make sense to integrate gitoxide in the
future as APIs improve, but only as a complete replacement of git2 and
not side by side

* Remove gix from josh-core
* I kept gix in proxy since it still provides nice Rust APIs for repo
init and config manipulation, however...
* ...I updated the version of gix to latest and disabled all extra
features, which reduced dependencies a lot