Fix slow push
To determine which files are part of a filtered subset,
regardless of their potentially changed path a trick is used:
The filter is applied and then the inverted filter is applied
to the result.
The implementation of that trick was to simply chain the filter
with the inverted one. However this caused flatten() to create
a hughe intermediate representation (something like a cartesian product)
that then has to be simplified down again by optimize()/simplify().
While that worked, it was very computationally expensive making
push extremly slow in some cases.
Now instead of chaining apply() is called twice, avoiding the
huge intermediate representation and thus expensive optimization.

Change: fix-slow-push