Fix prefix_sort ordering (#1469)
prefix_sort was using a sorting function that required, since we updated
rust, a total ordering.
In fact prefix_sort should reorder alphabetically, while preserving the
order of filters that share a prefix as input or output directory.

This commits replaces the function so that it does so, preventing
crashes in cases where the ordering was revealed to not be total.

Co-authored-by: Louis-Marie Givel <louis-marie.givel@accenture.com>