Fix chain(prefix(a/b),subdir(a))
The implementation here was wrong in that it assumed
subdir would be deeper than prefix.
Strangely the bug never manifested except now in a
scenario of building filters programatically.
I could not get the bug to reproduce by writing
a filter spec, so the regression test needs to be
a rust test.

Change: fix-chain-bug