Fix UI
The recent upgrade to hyper 1.0 seems to have broken the UI.
This commit fixes the issue with hyper_staticfile.
It also changes the return types from a fixed Response<Full<Bytes>> to a
custom JoshResponse type that allows us to deal with streaming responses
instead of always buffering them.
This type and some helper functions are defined in a new
hyper_integration.rs file