Add explicit options for orphan handling in push
If a push introduces new orphan history, there is some
ambiguity what that means. It could be a synthetic workspace
merge commit, an attempt to import an entirely new subtree
or just a new history that has been added to an existing
subtree.
Previously the "workspace edit" case was assumed and
"-p merge" could be used in the "add new subtree" case,
leaving no option for the "new history in existing subtree"
case.
Instead of guessing this change makes the push fail by default
and asks the user to be explicit in what to do.

Also change `merge_base_many` to `merge_base_octopus` as this
is the proper function to check for a common ancestor of multiple
commits according to git2 docs.