Use "populate" instead of "transpose" for pin filter
We didn't realize that we can use those existing functions instead
of creating a new one.
Going via pathstree() and populate() might not be the absolute best possible
route, but those use memoization and should be much faster for repeated
application that the transpose() which always needs to look at all paths.